Moved top bar to seperate folder
